Move files to AWS and slighty clean repo
