+++ archive/recommendation/mlops/model_registry.py
@@ -1,0 +1,4 @@
+import mlflow
+
+mlflow.set_tracking_uri("sqlite:///mlflow.db")
+mlflow.set_experiment("learnmate_recommendations")
