--- data/user_preferences.json
@@ -1,1 +1,0 @@
-{}
